requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.1
requests==2.31.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
